Moved loernwerkError to model
